implement java nested type overrided serialization

+++ java/fory-core/src/test/java/org/apache/fory/data/AllUnsignedFields.java
@@ -31,12 +31,12 @@ import org.apache.fory.config.LongEncoding;
 
 /** Test class with both unsigned integer scalar and array fields. */
 public class AllUnsignedFields {
-  @Uint8Type public byte u8;
+  @Uint8Type public int u8;
 
-  @Uint16Type public short u16;
+  @Uint16Type public int u16;
 
   @Uint32Type(compress = false)
-  public int u32;
+  public long u32;
 
   @Uint64Type(encoding = LongEncoding.FIXED)
   public long u64;
